fix: path of setTagsToProfile method
