+++ src/modules/profile/profile.module.ts
@@ -3,9 +3,10 @@ import { ProfileController } from './profile.controller';
 import { ProfileService } from './profile.service';
 import { ProfileRepository } from './profile.repository';
 import { PrismaModule } from '~/shared/adapters/prisma/prisma.module';
+import { TagModule } from '../tag/tag.module';
 
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, TagModule],
   controllers: [ProfileController],
   providers: [ProfileService, ProfileRepository],
   exports: [ProfileService, ProfileRepository],
