+++ src/modules/profile/profile.repository.ts
@@ -55,4 +55,13 @@ export class ProfileRepository {
       },
     });
   }
+
+  async setTagsToProfile(profileId: string, tags: string[]) {
+    return await this.prismaService.profile.update({
+      where: { id: profileId },
+      data: {
+        tags: { connect: tags.map((tag) => ({ id: tag })) },
+      },
+    });
+  }
 }
