+++ src/modules/tag/tag.controller.ts
@@ -1,8 +1,5 @@
-import { Body, Controller, Get, Post, Req } from '@nestjs/common';
+import { Controller, Get } from '@nestjs/common';
 import { TagService } from './tag.service';
-import { ISignedRequest } from '../auth/interfaces/signed-request.interface';
-import { ArrayLengthPipe } from '../../shared/utils/array-length.util';
-import { ArrayDuplicatedIndexPipe } from '../../shared/utils/array-duplicated-index.util';
 import { IGetTagsNames } from './interface/get-tags-names.interface';
 
 @Controller('tags')
@@ -12,13 +9,5 @@ export class TagController {
   @Get()
   async getTags(): Promise<IGetTagsNames[]> {
     return await this.tagService.getTagsNames();
-  }
-
-  @Post()
-  async setTagsToProfile(
-    @Req() req: ISignedRequest,
-    @Body(new ArrayLengthPipe(1, 5), ArrayDuplicatedIndexPipe) tags: string[],
-  ): Promise<void> {
-    await this.tagService.setTagsToProfile(req.user.profileId, tags);
   }
 }
